Provide default padding/switch behavior and overrides
If there are enough digits to show on all lights, it's nicer to not have
any blinking between invocations. Otherwise, it's needed so the start of
each digit and the entire sequence can be seen easily.

Make the default behavior to have the blinking if necessary but otherwise
disable it. Provide options so that the blinking can be forced if desired.