Implement set_light_optimized to avoid sending unnecessary bridge cmds
For programs which want to update light state frequently with dynamic
commands but save as much Zigbee bandwidth as possible. It saves the caller
the work of having to track the light state to avoid setting parameters
redundantly.

Also, apply a workaround to Bridge.set_light crashing if an empty parameter
list is specified, which is a distinct possibility when using
set_light_optimized