Fix mutual fields not updating while light off
The condition that avoided sending update commands to the bridge when the
light was off was also preventing field updates. This condition really
isn't necessary; it doesn't hurt to attempt to change a light that's off.
The main problem was the phue library dumping garbage onto the curses
UI when the bridge returned failure, but the logger configuration code in
__init__ solves that.