Fix ext. Kelvin field jumping to nearest mired when toggling light power
This requires refactoring do_field_update into separate methods for
recalculating the fields and sending updates to the lights. Then when the
power-on code tries to resend the parameters, it can call just the light
command sending code and not field recalculation (which is unnecessary
since the field values don't actually change).