Try to make light restore more reliable in weird situations
This is just a guess. The light states got messed up once while lights
became unreachable several times in short succession. Maybe the bridge got
confused and reported a bad status, and maybe the program shouldn't try to
rewrite it right after a light was restored in case the bridge isn't up to
date yet.