Add more robust handling of bridge connection failures to ExtendedBridge
This should allow requests to the Hue bridge to be auto-retried a few times
if an error occurs before actually bailing with an exception.
(My bridge apparently crashes and reboots itself on a regular basis, and I
got tired of it taking down all my scripts with it.)

Also, remove the restore_light_states_retry() method because the new logic
renders it redundant.