Add ability to temporarily disable Hue power-failure startup mode
This commit adds the new functionality to the fading/chasing_colors
and coded_digits/clock/stopwatch programs.

Hue lights can be configured to remember their previous state in case
a power loss occurs and return to it when power is restored. This
requires each light to periodically write its current state to
internal EEPROM while it's operating. Since these effects regularly
change the light state, this means EEPROM might be written to often
the entire time the effects run. In case this constant writing could
be detrimental to the EEPROM's cycle life, this commit adds an option
for temporarily switching the lights' startup mode to a fixed color
for the duration of execution, if the current mode saves the light's
state, and returning it to its original setting on exit.