Don't use set_light_optimized() where it's not really needed that badly
Although it theoretically makes Zigbee network usage more efficient,
it has the significant downside that if the lights lose power or are
otherwise changed my another app, the change will never be detected
and the lights could remain in an incorrect state for an
indeterminable period. So it's best not to use it unless the
performance benefit is worth the risk (chasing_colors is the only such
case I can think of right now).